refactor: Move Export Data to separate component
ExportData function is now imported to tabselector instead of
observation page.
