+++ src/components/deploymentLayout/TabSelector.tsx
@@ -1,9 +1,9 @@
 import { Link } from "@tanstack/react-router";
 
+import { ExportData } from "@/components/ExportData";
 import { buttonVariants } from "@/components/ui/button-variants";
 import { cn } from "@/lib/utils";
 import type { FileRoutesByTo } from "@/routeTree.gen";
-
 const tabs: { label: string; route: keyof FileRoutesByTo }[] = [
   { label: "Overview", route: "/deployment/$deploymentId/overview" },
   { label: "Analytics", route: "/deployment/$deploymentId/analytics" },
@@ -13,25 +13,30 @@ const tabs: { label: string; route: keyof FileRoutesByTo }[] = [
 
 function TabSelector() {
   return (
-    <nav className="bg-popover py-3">
-      <ul className="flex list-none gap-2 px-6">
-        {tabs.map((tab) => (
-          <li key={tab.route} className="flex">
-            <Link
-              to={tab.route}
-              params={(prev) => prev}
-              search={(prev) => prev}
-              className={cn(
-                buttonVariants({ variant: "secondary", size: "lg" }),
-                "tracking-wide uppercase [&.active]:bg-primary [&.active]:text-primary-foreground",
-              )}
-            >
-              {tab.label}
-            </Link>
-          </li>
-        ))}
-      </ul>
-    </nav>
+    <div className="flex flex-row bg-popover py-3">
+      <nav>
+        <ul className="flex flex-auto list-none gap-2 px-6">
+          {tabs.map((tab) => (
+            <li key={tab.route} className="flex">
+              <Link
+                to={tab.route}
+                params={(prev) => prev}
+                search={(prev) => prev}
+                className={cn(
+                  buttonVariants({ variant: "secondary", size: "lg" }),
+                  "tracking-wide uppercase [&.active]:bg-primary [&.active]:text-primary-foreground",
+                )}
+              >
+                {tab.label}
+              </Link>
+            </li>
+          ))}
+        </ul>
+      </nav>
+      <span className="flex flex-auto justify-end pr-8" aria-label="Export data button">
+        <ExportData />
+      </span>
+    </div>
   );
 }
 export { TabSelector };
